+++ client/doublezero/src/command/connect.rs
@@ -583,7 +583,7 @@ impl ProvisioningCliCommand {
                         spinner.println("❌ Error creating user");
                         spinner.println(format!("\n{}: {:?}\n", "Error", e));
 
-                        Err(eyre::eyre!("Error creating user"))
+                        Err(eyre::eyre!("Error creating user: {e:?}"))
                     }
                 }
             }?,
@@ -746,7 +746,9 @@ impl ProvisioningCliCommand {
                             Err(e) => {
                                 spinner.println("❌ Error adding publisher subscription");
                                 spinner.println(format!("\n{}: {:?}\n", "Error", e));
-                                eyre::bail!("Error adding publisher subscription to existing user");
+                                eyre::bail!(
+                                    "Error adding publisher subscription to existing user: {e:?}"
+                                );
                             }
                         }
                     }
@@ -776,7 +778,7 @@ impl ProvisioningCliCommand {
                                 spinner.println("❌ Error adding subscriber subscription");
                                 spinner.println(format!("\n{}: {:?}\n", "Error", e));
                                 eyre::bail!(
-                                    "Error adding subscriber subscription to existing user"
+                                    "Error adding subscriber subscription to existing user: {e:?}"
                                 );
                             }
                         }
@@ -829,7 +831,7 @@ impl ProvisioningCliCommand {
                     Err(e) => {
                         spinner.println("❌ Error creating user");
                         spinner.println(format!("\n{}: {:?}\n", "Error", e));
-                        return Err(eyre::eyre!("Error creating user"));
+                        return Err(eyre::eyre!("Error creating user: {e:?}"));
                     }
                 };