+++ smartcontract/sdk/rs/src/commands/multicastgroup/suspend.rs
@@ -16,7 +16,7 @@ impl SuspendMulticastGroupCommand {
             .execute(client)
             .map_err(|_err| eyre::eyre!("Globalstate not initialized"))?;
 
-        client.execute_transaction(
+        client.execute_authorized_transaction(
             DoubleZeroInstruction::SuspendMulticastGroup(MulticastGroupSuspendArgs {}),
             vec![
                 AccountMeta::new(self.pubkey, false),
@@ -48,7 +48,7 @@ mod tests {
         let (pda_pubkey, _) = get_location_pda(&client.get_program_id(), 1);
 
         client
-            .expect_execute_transaction()
+            .expect_execute_authorized_transaction()
             .with(
                 predicate::eq(DoubleZeroInstruction::SuspendMulticastGroup(
                     MulticastGroupSuspendArgs {},